Fixed #1: Added Cache on disk
This introduces a local filesystem cache for remote content in order to save download resources (time , bandwidth).

The Cache is _very flexible_ and works as follows:

- `--cache-dir` cli arg specifies a local folder to hold the cache.
- A policy must be present at `policy.yaml` in that folder ; otherwise, a default one is added
- Policy specifies which content should be allowed in the cache as well as constraints.
- if `--cache-dir` is not specified, a default `disabled` cache policy is used
- so from the code's POV, there's always a cache
- cache policy is checked on start
- then a Cache is created for the `cache_dir` and that policy
- the cache is then walked to find out about existing entries
- Cache content can be shown on the console with `--show-cache`
- Cache policy is applied: entries not confirming the policy anymore are removed
- ??? Outdated entries are removed from the cache (online has changed)
- Content URLs/Images are then checked. We know which oh them are in the cache.
- Those not in cache are added to a `candidates` list
- Policy is then applied to the candidates list. those not matching policy are dropped ; those matching are marked for insertion. It's possible that existing entries gets pushed out (removed) to leave room for new ones.
- Download then happens ; checking first for cache, copying in this case. If not in cache but should be, copied to cache after download

Reminder: there are 2 types of content in image_creator:
- File (base image and all other downloads)
- OCIImage (docker images)

The cache-allowance algorithm is in `cache.manager.get_eviction_for`.
It's probably easier to understand how it works by looking at the code than reading a description of it.
The basic idea is that entries gets in unconditionnaly ; either when reading fs on discovery or by checking content URLs.
Then the set of entries/candidates is filtered out by testing each entry in a specific order (according to Eviction Policy):
- for individual constraints like `ignore` flag or `max_age` or source protocol (we dont cache `file://`)
- for policy-level constraints like `max_size` and `max_num`.

There are several levels of embedded policies ; tested from deepest to highest:
- SubPolicy filters
- SubPolicies
- MainPolicy

The Policy documentation is pretty straightforward I believe as are outcomes (see README)

In addition to those policy-related evictions, in-cache entries are tested for outdacy: online URL is checked and it its digest has changed, we consider it outdated and evict it.
This allows having permalinks in the project YAML without worrying much.

That's why we don't cache `file://`. If *source* URI must be present to check outdacy and the source is already local, there's no point in using another local copy of the file.
Note that if the server (for files) doesnt supply an ETag not Content-Length+Modified Date, we have no way of knowing its *version* and consider is always outdated

An important tech point is that we need to store metadata about each files in the Cache. I chose to store them on the filesystem's extended attributes.
This seems like a good choice: it's invisible, it's strongly tied to each file, it doesn't require another file or DB.
It does require a compatible FS (ext4 with user_xattr for instance) but that has the default on Debian/Ubuntu for ages.
We'll see hot that stands in the future. If it becomes problematic ; we'd still be able to switch to a file-backed DB easily.

- Non-entry files in the cache are ignored. A file without a `source` xattr is considered foreign.
- It means that it's not possible to add files manually to the cache. Or one would need to set all the xattr manually.
- sys admins can query/manage the cache by using the cache-related options in combination with `--check` that stops after checks.
  - `--cache-dir xxx --show-cache --check`: prints the content of the cache
  - entries can be manually removed from cache by just removing the files
- Not-cache-related change: rewrote the Config parser/feedback using the method I used for the policy parsing: using self-validating dataclasses. Removes a lot of boilerplate. More readable.

- What should be the default policy? I've set `enabled=True, max_size=10GiB` for now.
- Should we jut not cache files without digest?