Added sizes checking feature
Before starting the main process of downloading and copying content into the image,
it is important to check that size constraints are met:
- size of content to go into the image (if a size was specified) will fit
- image size is under the maximum image size (if set ; new feature to set worker limits)
- disk space is available to process it all:
  - download (and maybe extract) base image
  - expand base image in target directory
  - download temp-files (those that needs to be extracted) into build-dir
  - download to-cache file into cache-dir

In order to calculate image size properly, the code needs to know the size that the
root filesystem in the base image occupies.

This information is missing and, from the base image file, can only be obtained after
it has been downloaded, extracted (if xz) and attached.
Problem being: we want to known before all download tasks whether the image request
can be satisfied or not.

Options to solve this:

We could defer the size computation step to after the base image has been retrieved
and attached.
It doesn't require any other input but it creates a possibility of failure if retrieving
the base image is not possible, due to filesize for instance.
We'd want to check that first and thus it would create two check steps reducing the
advantage of a single *all-good or failed* single step check.

That's the simplest and most straightforward option. Following other similar requirements
(for OCI images for instance), the base rootfs size would be mandatory.
Should the given value be incorrect (lower than reality), process would fail at attaching step.

This requires another to-be-known information while the previous behavior
of a single version or file URL was very convenient.

**that's what's currently implemented**

We could imagine appending the root fs size when building the base image

- to the .img file at a specific offset (end of file) but it would not be explicit, may break ability to just flash the base image directly
- to the .xz file as another file but it would be unexpected and not pigen-like which only compresses the .img file. Also, we want to support uncompressed img as well.

So, very hackish

When releasing a base image, we could upload a companion file (json? text?) that would have the same filename but with a dedicated suffix.
In conjunction with option B, if not set in image.yaml, code would try that companion file URL and use it if found.
This would mean a friction-less use for all *our releases* with the ability to input it manually in absence of the companion.

If requesting a rootfs size, we could consider it not an information but a request.
In this case, we'd not resize the third partition but delete it, resize the root partition
and then recreate the third partition.

Such a feature might be handy some day (given needs evolve differently than base) but
it's additional work for something that's not needed at this point.

- Only the first check calculating required image size and defining a margin would be shareable with higher-level tools.
- Added a new `--max-size` option allowing out-of-config bound to match resources (max size creatable by worker for instance)
- `base` config change from single string to dict with `source` and `rootfs_size`