Issue #21: Removed support for shrink option
Recent introduction of --shrink option lead to filesystem corruption on
last partition. Given the shrink option introduction was questioned
and done to satisfy an edge case “because it was a no brainer to add”,
I've decided to remove it altogether.

In default (`auto`) mode, image-creator computes all sizes before resize
(growing) the base-image's filesystem.
This allows to resize once, place all files in it and go:
an image file as large as its content.
In this case, the filesystem in the image is the size of the content.
A script run on first start on device resizes both the partition
and the filesystem so user still benefits from all size
of its sdcard.

It's also possible to request an arbitrary output image size (say 16GB).
In that case, the filesystem and the image file size are still accorded
but the file might include ton of void: if content is 10GB but you
request 128GB, you'll have a 128GB to download/store for nothing.
Use case for this is unclear but we had it in kiwix-hotspot because
we were not resizing partition and filesystem on start (exfat)

`--shrink` would be used only for that second use case so that users
would still get a smaller image file: the image still contains a 128GB
filesystem but the file itself is fit to the content (10GB).
That's really helpful for downloads and storage.

The problem is that shrink doesn't play well with ext4 because ext4
would spread data all around the place (it's fragmented).
We did not have this problem in kiwix-hotspot because
we were using exfat which apparently wrote contiguously.

Fixing shrink in image-creator would imply using resize2fs to shrink
the filesystem as well (and not just the file) which completely defeats
the purpose of the option: it becomes similar to `auto` mode.