Fixed #24 #25: Added `keep_identified_versions` Cache Policy Property
Requests the cache to keep only x versions (latests first) of a single entry.
For OCI Image entries, the version is compared using the tag.

For Files, the version is compared using a filename (on-disk) pattern: `_YYYY-MM.zim`.

This automatically removes any ~older version once a new one is available (and reaches the threshold)